+++ Sources/SwiftyModel/Query.swift
@@ -36,10 +36,6 @@ extension Repository {
     }
 }
 
-typealias ToOneRelation<T: EntityModel, Directionality: DirectionalityProtocol, Constraint: ConstraintsProtocol> = Relation<T, Directionality, Relations.ToOne, Constraint>
-
-typealias ToManyRelation<T: EntityModel, Directionality: DirectionalityProtocol, Constraint: ConstraintsProtocol> = Relation<T, Directionality, Relations.ToMany, Constraint>
-
 extension Query {
     
     func related<Child, Directionality, Constraints>(
